Moved patches related to Game1 and InstanceGame to their own class
