+++ stardew-access/Patches/InstanceGamePatch.cs
@@ -1,0 +1,11 @@
+namespace stardew_access.Patches
+{
+    internal class InstanceGamePatch
+    {
+        internal static void ExitPatch()
+        {
+            if (MainClass.ScreenReader != null)
+                MainClass.ScreenReader.CloseScreenReader();
+        }
+    }
+}
