--- stardew-access/Patches/QuestPatches.cs
@@ -1,11 +1,0 @@
-﻿using Microsoft.Xna.Framework.Graphics;
-using StardewValley;
-using StardewValley.Menus;
-using StardewValley.Quests;
-
-namespace stardew_access.Patches
-{
-    internal class QuestPatches
-    {
-    }
-}
